Update README.md
Ben 10 is my favourite cartoon character. Ben, while spending his vacation at his grandparent's house, finds a watch called Omnitrix. The watch permanently attaches to him and gives him superpowers. He, with the help of the watch, can turn to ten alien forms.
Dragon Ball tells the tale of a young warrior by the name of Son Goku, a young peculiar boy with a tail who embarks on a quest to become stronger and learns of the Dragon Balls, when, once all 7 are gathered, grant any wish of choice.
Vampire brothers, Stefan and Damon Salvatore, battle for the affection of selfless teenager, Elena Gilbert. Stefan is the good and kind-hearted brother that feeds on animal blood to keep from killing anybody, and Damon is the selfish and dangerous older brother that feeds on human blood carelessly.